Undo label component re-organization
